Refactor JSON serialising code and add support for Arrow
